Seperated app from server
