+++ index.js
@@ -1,29 +1,8 @@
-const express = require("express");
-const cookieSession = require("cookie-session");
-const app = express();
+const app = require("./app");
 const port = 3000;
 
 const db = require("./src/models");
 const dbConfig = require("./src/config/db.config");
-const authConfig = require("./src/config/auth.config");
-
-app.use(express.json());
-app.use(
-  express.urlencoded({
-    extended: true,
-  })
-);
-app.use(
-  cookieSession({
-    name: "chat-app-session",
-    secret: authConfig.cookieSecret,
-    httpOnly: true,
-  })
-);
-app.use(function (req, res, next) {
-  res.header("Access-Control-Allow-Headers", "Origin, Content-Type, Accept");
-  next();
-});
 
 // DB Connection
 db.mongoose
@@ -33,19 +12,7 @@ db.mongoose
 
 db.init();
 
-// Test route
-app.get("/api", (req, res) => {
-  res.json({ message: "ok" });
-});
-
-// Import routes
-const authRouter = require("./src/routes/auth.routes");
-const testRouter = require("./src/routes/test.routes");
-
-// Use routes
-app.use("/api/auth", authRouter);
-app.use("/api/test", testRouter);
-
+// Express app
 app.listen(port, () => {
   console.log(`Chat-app API listening on port ${port}!`);
 });
